+++ versioned/storage/cache/src/main/java/org/projectnessie/versioned/storage/cache/PersistCaches.java
@@ -15,11 +15,13 @@
  */
 package org.projectnessie.versioned.storage.cache;
 
+import io.micrometer.core.instrument.MeterRegistry;
+
 public final class PersistCaches {
   private PersistCaches() {}
 
   /** Produces a {@link CacheBackend} with the given maximum capacity. */
-  public static CacheBackend newBackend(long capacity) {
-    return CaffeineCacheBackend.builder().capacity(capacity).build();
+  public static CacheBackend newBackend(long capacityMb, MeterRegistry meterRegistry) {
+    return CaffeineCacheBackend.builder().capacity(capacityMb).meterRegistry(meterRegistry).build();
   }
 }